examples/basic/requirements.txt
django-model-utils
jinja2
django-countries
